extract trimMessage out of server.js
